mapper 오류 수정
