+++ frontend/eco-insight/src/App.jsx
@@ -11,9 +11,6 @@ import OkWithdrawal from "./components/MyPage/Withdrawal/OkWithdrawal";
 import CheckWithdrawal from "./components/MyPage/Withdrawal/CheckWithdrawal";
 import EditProfile from "./components/MyPage/EditProfile";
 import Footer from "./components/Common/Footer/Footer";
-
-import WritePostPage from "./components/Board/WritePostPage";
-
 import UserLayout from "./components/Layout/UserLayout";
 import AdminRoute from "./components/RouteGuard/AdminRoute";
 import AdminLogin from "./pages/Admin/AdminLogin/AdminLogin.jsx";
@@ -44,8 +41,8 @@ import NoticeBoard from "./components/Board/Notice/NoticeBoard.jsx";
 import NoticeBoardDetail from "./components/Board/Notice/NoticeBoardDetail.jsx";
 import NoticeBoardWrite from "./components/Board/Notice/NoticeBoardWrite.jsx";
 import NoticeWrite from "./pages/Admin/NoticeBoardManagementPage/NoticeWrite.jsx";
-import Notice from "./components/Board/Notice/NoticeBoard.jsx"
-
+import Notice from "./components/Board/Notice/NoticeBoard.jsx";
+import CommunityWritePage from "./components/Board/Community/CommunityWritePage.jsx";
 
 function App() {
   return (
@@ -58,25 +55,22 @@ function App() {
           <Route path="/board/cert" element={<AuthListPage />} />
           <Route path="/board/cert/:no" element={<AuthBoardDetail />} />
           <Route path="/board/cert:type" element={<AuthBoardWritePage />} />
-          <Route path="/write/:type" element={<WritePostPage />} />
+          <Route path="/write/:type" element={<CommunityWritePage />} />
           <Route path="/post/:id" element={<CommunityBoardDetail />} />
           <Route path="/dashboard" element={<MainDashBoard />} />
-
           <Route path="/frequencyAskPage" element={<FrequencyAskPage />} />
           <Route path="/privateAskPage" element={<PrivateAskPage />} />
           <Route path="/find-id/result-id" element={<ResultId />} />
-
-        {/* 로그인 유저 전용 */}
-        <Route element={<UserRoute />}>
-          <Route path="/mypage"          element={<MyPage />} />
-          <Route path="/myposts"         element={<Myposts />} />
-          <Route path="/editprofile"     element={<EditProfile />} />
-          <Route path="/changepassword"  element={<ChangePassword />} />
-          <Route path="/withdrawal/check" element={<CheckWithdrawal />} />
-          <Route path="/withdrawal/form"  element={<WithdrawalForm />} />
-          <Route path="/withdrawal/ok"    element={<OkWithdrawal />} />
-        </Route>
-
+          {/* 로그인 유저 전용 */}
+          <Route element={<UserRoute />}>
+            <Route path="/mypage" element={<MyPage />} />
+            <Route path="/myposts" element={<Myposts />} />
+            <Route path="/editprofile" element={<EditProfile />} />
+            <Route path="/changepassword" element={<ChangePassword />} />
+            <Route path="/withdrawal/check" element={<CheckWithdrawal />} />
+            <Route path="/withdrawal/form" element={<WithdrawalForm />} />
+            <Route path="/withdrawal/ok" element={<OkWithdrawal />} />
+          </Route>
           <Route path="/notice" element={<Notice />} />
           <Route path="/board/notice/:id" element={<NoticeBoardDetail />} />
           <Route path="/frequencyAskPage" element={<FrequencyAskPage />} />
@@ -145,14 +139,13 @@ function App() {
           />{" "}
           {/* findpassword,changepassword 추가 예정 */}
           <Route path="/changepassword" element={<ChangePassword />} />
-
         </Route>
 
         <Route path="/login" element={<Login />} />
         <Route path="/signup" element={<SignUp />} />
         <Route path="/findid" element={<FindId />} />
         <Route path="/findpassword" element={<FindPasswordPage />} />
-        <Route path="/findPassword/resetpassword" element={<ResetPassword />}/> 
+        <Route path="/findPassword/resetpassword" element={<ResetPassword />} />
         {/* 관리자 전용 */}
         <Route path="/admin/login" element={<AdminLogin />} />
 
@@ -167,9 +160,18 @@ function App() {
           <Route path="dashboard" element={<AdminDashBoard />} />
           <Route path="point-manage" element={<PointManagementPage />} />
           <Route path="account-manage" element={<AccountManagementPage />} />
-          <Route path="authboard-manage" element={<AuthBoardManagementPage />} />
-          <Route path="communityboard-manage" element={<CommunityBoardManagementPage />} />
-          <Route path="noticeboard-manage" element={<NoticeBoardManagementPage />} />
+          <Route
+            path="authboard-manage"
+            element={<AuthBoardManagementPage />}
+          />
+          <Route
+            path="communityboard-manage"
+            element={<CommunityBoardManagementPage />}
+          />
+          <Route
+            path="noticeboard-manage"
+            element={<NoticeBoardManagementPage />}
+          />
 
           {/*<Route path="notice-write" element={<NoticeWrite />} />*/}
           <Route path="notice-write" element={<NoticeBoardWrite />} />
@@ -186,7 +188,6 @@ function App() {
             element={<NoticeBoardManagementPage />}
           />
           <Route path="notice-write" element={<NoticeWrite />} />
-
         </Route>
       </Routes>
     </AuthProvider>
